busted google auth
Looking for a fresh set of eyes to take a look at this. It should work, but it doesn't. The main issue I've been having trouble with was the redirectUri which just doesn't work for some reason. I made this on my illinois account tho, so I'll send the relevant links on discord. It was using the local ip address as the redirectUri so I tried hardcoing the redirect uri in login.tsx in the useAuthRequest, but it doesn't work at the very end...